feat: regions now check if visted before using playerprefs

+++ Assets/Scripts/Region/Region.cs
@@ -9,7 +9,7 @@ namespace mattordev.regions
         public string regionName;
         public BoxCollider regionCollider; // region collider.
         public AudioClip regionTransitionSFX; // sound effect to play when entering this region.
-        public float scoreBoostWhenEntering = 0f; // score boost when entering this region.
+        public int scoreBoostWhenEntering = 0; // score boost when entering this region.
 
         public enum weatherType
         {
